backend/requirements.txt
Django>=3.2.3,<3.3
djangorestframework>=3.12.4,<3.13
django-cors-headers>=3.10.1,<3.11
psycopg2-binary>=2.8.5,<2.9
gunicorn>=20.1.0,<20.2